requirements.txt
annotated-types==0.7.0
anyio==4.8.0
asttokens==3.0.0
attrs==25.1.0
backcall==0.2.0
beautifulsoup4==4.13.3
bleach==6.2.0
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
cryptography==44.0.2
dbrider==1.0.3
decorator==5.2.1
defusedxml==0.7.1
dnspython==2.7.0
docopt==0.6.2
email_validator==2.2.0
exceptiongroup==1.2.2
executing==2.2.0
Faker==36.1.1
fastapi==0.115.11
fastjsonschema==2.21.1
greenlet==3.1.1
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
ipython==8.12.3
jedi==0.19.2
Jinja2==3.1.6
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter_client==8.6.3
jupyter_core==5.7.2
jupyterlab_pygments==0.3.0
MarkupSafe==3.0.2
matplotlib-inline==0.1.7
mistune==3.1.2
nbclient==0.10.2
nbconvert==7.16.6
nbformat==5.10.4
packaging==24.2
pandocfilters==1.5.1
parso==0.8.4
passlib==1.7.4
peewee==3.14.10
pexpect==4.9.0
pickleshare==0.7.5
platformdirs==4.3.6
playhouse==0.1.0
prompt_toolkit==3.0.50
psycopg==3.2.5
psycopg2-binary==2.9.10
ptyprocess==0.7.0
pure_eval==0.2.3
pycparser==2.22
pydantic==2.10.6
pydantic_core==2.27.2
Pygments==2.19.1
PyJWT==2.10.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
PyYAML==6.0
pyzmq==26.2.1
referencing==0.36.2
requests==2.32.3
rpds-py==0.23.1
six==1.17.0
sniffio==1.3.1
soupsieve==2.6
SQLAlchemy==2.0.38
stack-data==0.6.3
starlette==0.46.0
tinycss2==1.4.0
tornado==6.4.2
traitlets==5.14.3
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.3.0
uvicorn==0.34.0
wcwidth==0.2.13
webencodings==0.5.1
yarg==0.1.9